Update RegisterMenuController, TradeMenuController & CreateNewGameMenuController.
