+++ src/controller/CreateNewGameMenuController.java
@@ -1,0 +1,5 @@
+package controller;
+
+public class CreateNewGameMenuController {
+
+}
