+++ src/controller/RegisterMenuController.java
@@ -1,4 +1,22 @@
 package controller;
 
+import java.util.regex.Matcher;
+
 public class RegisterMenuController {
+    public static String register(Matcher matcher){
+        return "...";
+    }
+
+    public static String randomQuestion(Matcher matcher){
+        return "...";
+    }
+
+    public static String randomPassword(Matcher matcher){
+        return "...";
+    }
+
+    public static String randomSlogan(Matcher matcher){
+        return "...";
+    }
+
 }
