+++ src/controller/TradeMenuController.java
@@ -1,4 +1,22 @@
 package controller;
 
+import java.util.regex.Matcher;
+
 public class TradeMenuController {
+    public static String tradeRequest(Matcher matcher){
+        return "...";
+    }
+
+    public static String tradeList(Matcher matcher){
+        return "...";
+    }
+
+    public static String tradeAccept(Matcher matcher){
+        return "...";
+    }
+
+    public static String tradeHistory(Matcher matcher){
+        return "...";
+    }
+
 }
